gen_flow/requirements/base.txt
# SHA1:62e78eeccf109aae8716f4a042b8f5e35472e086
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
asgiref==3.8.1
    # via
    #   channels
    #   daphne
    #   django
    #   django-cors-headers
attrs==25.1.0
    # via
    #   jsonschema
    #   referencing
    #   service-identity
    #   twisted
autobahn==24.4.2
    # via daphne
automat==24.8.1
    # via twisted
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
channels[daphne]==4.2.0
    # via -r gen_flow/requirements/base.in
charset-normalizer==3.4.1
    # via requests
constantly==23.10.4
    # via twisted
cryptography==44.0.2
    # via
    #   autobahn
    #   pyjwt
    #   pyopenssl
    #   service-identity
daphne==4.1.2
    # via channels
defusedxml==0.7.1
    # via python3-openid
dj-pagination==2.5.0
    # via -r gen_flow/requirements/base.in
dj-rest-auth[with-social]==5.0.2
    # via -r gen_flow/requirements/base.in
django==5.1.6
    # via
    #   -r gen_flow/requirements/base.in
    #   channels
    #   dj-rest-auth
    #   django-allauth
    #   django-appconf
    #   django-compressor
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   drf-spectacular
django-allauth[saml]==0.57.2
    # via
    #   -r gen_flow/requirements/base.in
    #   dj-rest-auth
django-appconf==1.1.0
    # via django-compressor
django-compressor==4.5.1
    # via -r gen_flow/requirements/base.in
django-cors-headers==4.7.0
    # via -r gen_flow/requirements/base.in
django-filter==25.1
    # via -r gen_flow/requirements/base.in
djangorestframework==3.15.2
    # via
    #   -r gen_flow/requirements/base.in
    #   dj-rest-auth
    #   drf-spectacular
drf-spectacular==0.28.0
    # via -r gen_flow/requirements/base.in
gmpy2==2.2.1
    # via -r gen_flow/requirements/base.in
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.10
    # via
    #   hyperlink
    #   requests
    #   twisted
incremental==24.7.2
    # via twisted
inflection==0.5.1
    # via drf-spectacular
isodate==0.7.2
    # via python3-saml
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
lxml==5.3.1
    # via
    #   python3-saml
    #   xmlsec
oauthlib==3.2.2
    # via requests-oauthlib
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.1
    # via service-identity
pycparser==2.22
    # via cffi
pycryptodome==3.21.0
    # via -r gen_flow/requirements/base.in
pyjwt[crypto]==2.10.1
    # via django-allauth
pyopenssl==25.0.0
    # via twisted
python3-openid==3.2.0
    # via django-allauth
python3-saml==1.16.0
    # via django-allauth
pyyaml==6.0.2
    # via drf-spectacular
rcssmin==1.1.2
    # via django-compressor
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via django-allauth
rjsmin==1.2.2
    # via django-compressor
rpds-py==0.23.1
    # via
    #   jsonschema
    #   referencing
service-identity==24.2.0
    # via twisted
sqlparse==0.5.3
    # via django
tomli==2.2.1
    # via incremental
twisted[tls]==24.11.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.12.2
    # via
    #   asgiref
    #   pyopenssl
    #   referencing
    #   twisted
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.3.0
    # via requests
xmlsec==1.3.14
    # via python3-saml
zope-interface==7.2
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
setuptools==76.0.0
    # via
    #   autobahn
    #   incremental
    #   zope-interface
